Convert CDA data hooks from TS to JS
