--- lib/components/water-management/useCdaCatalogTS.ts
@@ -1,69 +1,0 @@
-import { UseQueryOptions, useQuery } from "@tanstack/react-query";
-import { buildRequest } from "./cda";
-
-// Return types generated by https://app.quicktype.io/
-interface CdaCatalogTS {
-  total: number;
-  entries: Entry[];
-  "page-size": number;
-}
-
-interface Entry {
-  office: string;
-  name: string;
-  units: string;
-  interval: string;
-  "interval-offset": number;
-  "time-zone": string;
-  extents: Extent[];
-}
-
-interface Extent {
-  "earliest-time": string;
-  "last-update": string;
-  "latest-time": string;
-}
-
-type GetCdaCatalogParams = {
-  office: string;
-  like: string;
-};
-
-const getCdaCatalogTS = async (
-  cdaParams: GetCdaCatalogParams,
-  cdaUrl?: string
-): Promise<CdaCatalogTS> => {
-  const paramString = new URLSearchParams(cdaParams).toString();
-  const url = buildRequest("/catalog/TIMESERIES", paramString, cdaUrl);
-  const response = await fetch(url, {
-    headers: {
-      accept: "application/json;version=2",
-    },
-  });
-  if (!response.ok) {
-    throw new Error("Error retrieving data.");
-  }
-  return response.json();
-};
-
-interface useCdaCatalogTSParams {
-  cdaParams: GetCdaCatalogParams;
-  cdaUrl?: string;
-  queryOptions?: Omit<UseQueryOptions<CdaCatalogTS>, "queryKey" | "queryFn">;
-}
-
-const useCdaCatalogTS = ({
-  cdaParams,
-  cdaUrl,
-  queryOptions,
-}: useCdaCatalogTSParams) =>
-  useQuery({
-    queryKey: ["cda", "catalog", cdaParams.like],
-    queryFn: async () => {
-      return getCdaCatalogTS(cdaParams, cdaUrl);
-    },
-    ...queryOptions,
-  });
-
-export { useCdaCatalogTS };
-export default useCdaCatalogTS;
