--- lib/components/water-management/useCdaLocation.ts
@@ -1,72 +1,0 @@
-import { UseQueryOptions, useQuery } from "@tanstack/react-query";
-import { buildRequest } from "./cda";
-
-// Return types generated by https://app.quicktype.io/
-interface CdaLocation {
-  "office-id": string;
-  name: string;
-  latitude: number;
-  longitude: number;
-  active: boolean;
-  "public-name": string;
-  "long-name": string;
-  "timezone-name": string;
-  "location-type": string;
-  "location-kind": string;
-  nation: string;
-  "state-initial": string;
-  "county-name": string;
-  "nearest-city": string;
-  "horizontal-datum": string;
-  "vertical-datum": string;
-  elevation: number;
-  "map-label": string;
-  "bounding-office-id": string;
-}
-
-type GetCdaLocationParams = {
-  office: string;
-  unit?: string;
-};
-
-const getCdaLocation = async (
-  location: string,
-  cdaParams: GetCdaLocationParams,
-  cdaUrl?: string
-): Promise<CdaLocation> => {
-  const paramString = new URLSearchParams(cdaParams).toString();
-  const url = buildRequest(`/locations/${location}`, paramString, cdaUrl);
-  const response = await fetch(url, {
-    headers: {
-      accept: "application/json;version=2",
-    },
-  });
-  if (!response.ok) {
-    throw new Error("Error retrieving data.");
-  }
-  return response.json();
-};
-
-interface useCdaLocationParams {
-  cdaParams: { location: string } & GetCdaLocationParams;
-  cdaUrl?: string;
-  queryOptions?: Omit<UseQueryOptions<CdaLocation>, "queryKey" | "queryFn">;
-}
-
-const useCdaLocation = ({
-  cdaParams,
-  cdaUrl,
-  queryOptions,
-}: useCdaLocationParams) => {
-  const { location, ...restParams } = cdaParams;
-  return useQuery({
-    queryKey: ["cda", "location", location],
-    queryFn: async () => {
-      return getCdaLocation(location, restParams, cdaUrl);
-    },
-    ...queryOptions,
-  });
-};
-
-export { useCdaLocation };
-export default useCdaLocation;
